added partial update functionality and other tweaks

+++ backend/src/main/java/com/kinsella/kinsellaBackend/repositories/IAccountsRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Collection;
 
 
-public interface IAccountsRepository extends JpaRepository<Accounts, Integer> {
+public interface IAccountsRepository extends JpaRepository<Accounts, Long> {
     Collection<Accounts> findByUserID(Long user);
 
     Accounts findByPhoneNumber(String phoneNumber);
